requirements.txt
# Web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Django ORM (standalone)
django==5.0.1
psycopg2-binary==2.9.9

# Authentication
pyjwt==2.8.0
passlib[bcrypt]==1.7.4

# Data validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Redis
redis==5.0.1

# HTTP client
httpx==0.26.0
requests==2.31.0

# Database
sqlparse==0.4.4

# YAML config
pyyaml==6.0.1

# Export
openpyxl==3.1.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Linting
ruff==0.1.14
mypy==1.8.0
